Adds Setup Support
This project is currently used in a way where all consuming projects
must git clone this to a tmp directory and copy it to the local
directory.  With python this isn't necessary and we can provide
access to this project in a path by adding it into a project's
requirements-test.txt file.

This patch adds a setup.py file that will copy the ci_harness and
associated task files into the bin of the virtualenv.  With these
files on the path, they no longer require the complexity of make
target "anchore-ci:".  The documentation has been updated to provide
examples of installation and remove references to some of the make
usage no longer needed.

The process of using pip to manage the installation of this project
is served well by using releases/tags.  Because this project does
not change frequently, this is not a difficult or time consuming
process and takes just a moment in the github UI.

Signed-off-by: Ryan Brady <ryan.brady@anchore.com>